# services/oid4vci/deploy/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: oid4vci API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8088
  description: Auto generated value
- url: http://0.0.0.0:8088
  description: Auto generated value
paths:
...
  /oid4vci/credential:
    post:
      tags:
      - Oidc 4 Vci Impl
      summary: OID4VCI Credential Endpoint
      description: Requests credential retrieval (cf. Sec. 7).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialRequest'
      responses:
        "200":
          description: OK
  /oid4vci/token:
    post:
      tags:
      - Oidc 4 Vci Impl
      summary: OID4VCI Token Endpoint
      description: Retrieves the token for the following credential issuance (cf.
        Sec. 6).
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
              - grant_type
              type: object
              properties:
                grant_type:
                  pattern: \S
                  type: string
                code:
                  type: string
                  nullable: true
                redirect_uri:
                  type: string
                  nullable: true
                client_id:
                  type: string
                  nullable: true
                pre-authorized_code:
                  type: string
                  nullable: true
                tx_code:
                  type: string
                  nullable: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Oid4VciTokenResponse'
components:
  schemas:
    AnyType:
      type: object
    AuthorizationDetails:
      required:
      - type
      - credential_identifiers
      type: object
      properties:
        type:
          type: string
        credential_identifiers:
          type: array
          items:
            type: string
...
    CredentialRequest:
      type: object
      properties:
        proof:
          type: object
          allOf:
          - $ref: '#/components/schemas/Proof'
          nullable: true
        credential_response_encryption:
          type: object
          allOf:
          - $ref: '#/components/schemas/CredentialResponseEncryption'
          nullable: true
    CredentialResponseEncryption:
      required:
      - jwk
      - alg
      - enc
      type: object
      properties:
        jwk:
          $ref: '#/components/schemas/ObjectNode'
        alg:
          type: string
        enc:
          type: string
...
    OauthTokenResponse:
      required:
      - access_token
      - token_type
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          format: int64
          type: integer
          nullable: true
        refresh_token:
          type: string
          nullable: true
        scope:
          type: string
          nullable: true
    ObjectNode:
      type: array
      items:
        $ref: '#/components/schemas/AnyType'
    Oid4VciTokenResponse:
      required:
      - oauth_token_response
      - authorization_details
      type: object
      properties:
        oauth_token_response:
          $ref: '#/components/schemas/OauthTokenResponse'
        c_nonce:
          type: string
          nullable: true
        c_nonce_expires_in:
          format: int64
          type: integer
          nullable: true
        authorization_details:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizationDetails'
...
    Proof:
      required:
      - proof_type
      type: object
      properties:
        proof_type:
          $ref: '#/components/schemas/ProofTypeEnum'
    ProofTypeEnum:
      enum:
      - jwt
      - cwt
      - ldp_vp
      type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT